Fix pipeline stuck (sometimes) when `replicas>1`
Fix pipeline getting stuck when the number of replicas of one step was
higher than 1. The pipeline was getting stuck because some of the
replicas didn't finish its execution because some other replicas were
consuming more than one of the `None`s (1 per replica) sent via the input queue to
notify to stop.